se reubica contenido de component con datos de alumno a conponent genrado para uso
