+++ src/app/components/Layout/toolbar/toolbar.component.html
@@ -5,7 +5,7 @@
         </div>
         <a class="navbar-brand" href="#">
             <img class="img-logo" src="https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQy7gbhTNgTryGj3Y5ug9FtBqKks08o-ONw85envpyTYPP2VBGAv3kkH2wawsXFC3drbnY&usqp=CAU" alt="">
-            Curso Angular
+            Cursos Online
           </a>
         <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarText" aria-controls="navbarText" aria-expanded="false" aria-label="Toggle navigation">
           <span class="navbar-toggler-icon"></span>
