+++ src/app/components/alumno/alumno.component.html
@@ -1,2 +1,96 @@
-<p>alumno works!</p>
-<h1>Lista de alumnos</h1>
+<section class="sec-tables">
+    <div class="row mt-5">
+        <div class="col-10 offset-1">
+            <div class="row mb-3">
+                <div class="col-md-12">
+                    <label for="">Seleccione tipo de busqueda</label>
+                </div>
+                <div class="col-md-6">
+                    <!-- <h4>Seleccione tipo de busqueda</h4>
+                    <mat-form-field appearance="fill">
+                    <mat-label>Favorite food</mat-label>
+                    <mat-select [(ngModel)]="selectedValue" name="food">
+                        <mat-option *ngFor="let food of foods" [value]="food.value">
+                        {{food.viewValue}}
+                        </mat-option>
+                    </mat-select> -
+                    </mat-form-field>-->
+                    <select class="custom-select" [(ngModel)]="opcionSeleccionado" (change)="capturar()">
+                        <option hidden >Seleccione Opcion deseada</option>
+                        <option [value]="1">Carrera</option>
+                        <option [value]="2">Institucion</option>
+                        <!-- <option [value]="3">Nombre</option> -->
+                    </select>
+                </div>
+                <div class="col-md-6 d-flex justify-content-center align-items-center ">
+                    <div class="row">
+                        <button mat-flat-button color="primary" (click)="addAlumno()"><mat-icon>person_add</mat-icon>&nbsp; Agregar Nuevo Alumno</button>
+                    </div>
+                </div>
+            </div>
+            <div class="row" *ngIf="filtroBusqueda">
+                <div class="col-md-6 ">
+                    <label class=" mt-2" for="">Ingrese {{valorFIltro | uppercase}}: </label>
+                    <div class="d-flex">
+                        <div>
+                            <input type="email" class="form-control" id="exampleInputEmail1" aria-describedby="emailHelp" placeholder="" [formControl]="textoDeInput" [ngModel]="valorInput" >
+                        </div>
+                        <div class="ml-3">
+                            <!-- <button  *ngIf='textoDeInput' class="btn btn-success" >Buscar</button> -->
+                            <button *ngIf='textoDeInput' mat-raised-button color="success" (click)="ValorBusqueda()"><mat-icon>search</mat-icon>&nbsp; Success</button>
+                            <button mat-raised-button color="warn" class="ml-3" (click)="nuevaBusqueda()"><mat-icon>reply</mat-icon>&nbsp; Regresar</button>
+                        </div>
+                    </div>
+                </div>
+            </div>
+            <br><br>
+            <div class="row p-2 ">
+                <table  mat-table  [dataSource]="dataSource"  class="mat-elevation-z8">                    
+                        <!-- Columna id -->
+                        <ng-container matColumnDef="id">
+                            <th appTheadTable mat-header-cell *matHeaderCellDef>{{ 'No.' | uppercase}} </th>
+                            <td mat-cell *matCellDef="let dataArreglo"> {{dataArreglo.id}} </td>
+                        </ng-container>
+                        <!-- Columna Nombre -->
+                        <ng-container matColumnDef="nombre">
+                            <th appTheadTable mat-header-cell *matHeaderCellDef> {{'Nombre Completo' | uppercase}} </th>
+                            <td mat-cell *matCellDef="let dataArreglo">{{ dataArreglo | PipeAlumno }}</td>
+                        </ng-container>
+                        <!-- Columna edad -->
+                        <ng-container matColumnDef="edad">
+                            <th appTheadTable mat-header-cell *matHeaderCellDef> {{ 'Edad' | uppercase}} </th>
+                            <td mat-cell *matCellDef="let dataArreglo"> {{dataArreglo.edad}} </td>
+                        </ng-container>
+                        <!-- Columna carrera -->
+                        <ng-container matColumnDef="carrera">
+                            <th appTheadTable mat-header-cell *matHeaderCellDef> {{ 'Carrera' | uppercase}} </th>
+                            <td mat-cell *matCellDef="let dataArreglo"> {{dataArreglo.carrera}} </td>
+                        </ng-container>
+                        <!-- Columna institucion -->
+                        <ng-container matColumnDef="institucion">
+                            <th appTheadTable mat-header-cell *matHeaderCellDef> {{ 'Institucion' | uppercase}} </th>
+                            <td mat-cell *matCellDef="let dataArreglo"> {{dataArreglo.institucion}} </td>
+                        </ng-container>
+                        <ng-container matColumnDef="star" stickyEnd>
+                            <th appTheadTable class="text-center" mat-header-cell *matHeaderCellDef aria-label="row actions">{{ 'Acciones' | uppercase}}</th>
+                            <td class="text-center" mat-cell *matCellDef="let element">
+                                <button class="m-1" mat-icon-button matTooltip="Eliminar Alumno" color="warn" (click)="deleteAlumno(element)">
+                                    <mat-icon>delete</mat-icon>
+                                </button>
+                                <button class="m-1" mat-icon-button color="primary" (click)="updateAlumno(element)"  matTooltip="Actualizar informacion">
+                                    <mat-icon>edit</mat-icon>
+                                </button>
+                                <!-- <button class="m-1" mat-icon-button color="primary" matTooltip="Mostrar Informacion">
+                                    <mat-icon>info</mat-icon>
+                                </button> -->
+                            </td>
+                        </ng-container>
+                        <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+                        <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
+                    </table>
+                </div>
+            </div>
+        </div>
+    <br><br>
+</section>
+
