+++ src/app/components/cursos/cursos.component.css
@@ -10,3 +10,4 @@
 .mat-card-actions{
     display: flex;
 }
+
