+++ src/app/components/cursos/cursos.component.html
@@ -3,16 +3,13 @@
   <div class="row pt-4">
     <h1>Seccion Cursos</h1>
   </div>
-  <div class="row">
+  <div class="row pb-4">
     <p> Más de 1000 Cursos Online Certificados en Crehana! Encuentra Cursos en linea y capacitaciones de todas las categorías para aprender a tu ritmo.</p>
     <br>
     <p>Descubre un amplio contenido de formación sin coste para impulsar tu negocio o carrera profesional. Aprende todo lo que necesites seleccionando módulos concretos o, si lo prefieres, haz un curso completo.</p>
   </div>
-  
 </div>
-
-
-
+<br>
 <div class="container">
 <div class="row pt-3">
       <mat-card *ngFor="let data of cursos">
@@ -21,9 +18,9 @@
           <mat-card-title>{{data.nombre}}</mat-card-title>
           <mat-card-subtitle>{{data.area}}</mat-card-subtitle>
         </mat-card-header>
-        <img mat-card-image src={{data.img}} alt="Photo of a Shiba Inu">
+        <img mat-card-image src={{data.img}} >
         <mat-card-content>
-          <p>
+          <p class="text-justify">
             {{data.descripcion}}
           </p>
         </mat-card-content>
